requirements.txt
streamlit==1.28.1
google-genai==0.3.0
